+++ OAuth/Controllers/Filters/ApiActionFilter.cs
@@ -10,7 +10,6 @@ namespace OAuth.Controllers.Filters
 {
     internal class ApiActionFilter : IAsyncActionFilter
     {
-        /// <inheritdoc/>
         public async Task OnActionExecutionAsync(ActionExecutingContext actionExecutingContext, ActionExecutionDelegate next)
         {
             if (CheckModelState(actionExecutingContext))
@@ -49,6 +48,7 @@ namespace OAuth.Controllers.Filters
                     ExceptionApiProblemDetails(actionExecutingContext, actionExecutedContext, tex);
 
                     return Task.FromCanceled(new CancellationToken(true));
+
                 case Exception ex:
                     actionExecutingContext.HttpContext.Response.StatusCode = StatusCodes.Status500InternalServerError;
                     ExceptionApiProblemDetails(actionExecutingContext, actionExecutedContext, ex);
@@ -58,7 +58,19 @@ namespace OAuth.Controllers.Filters
 
             switch (actionExecutedContext.Result)
             {
-                case NotFoundResult nfr:
+                case NotFoundObjectResult nfor:
+                    actionExecutedContext.HttpContext.Response.StatusCode = StatusCodes.Status404NotFound;
+                    AfterResultApiProblemDetails(actionExecutedContext, nfor.Value.ToString());
+
+                    return Task.FromCanceled(new CancellationToken(true));
+
+                case BadRequestObjectResult bror:
+                    actionExecutedContext.HttpContext.Response.StatusCode = StatusCodes.Status404NotFound;
+                    AfterResultApiProblemDetails(actionExecutedContext, bror.Value.ToString());
+
+                    return Task.FromCanceled(new CancellationToken(true));
+
+                case NotFoundResult _:
                     actionExecutedContext.HttpContext.Response.StatusCode = StatusCodes.Status404NotFound;
                     AfterResultApiProblemDetails(actionExecutedContext);
 
@@ -71,28 +83,31 @@ namespace OAuth.Controllers.Filters
         private void ValidateApiProblemDetails(ActionExecutingContext actionExecutingContext)
         {
             var apiProblemDetails = new ApiProblemDetails(actionExecutingContext.HttpContext, actionExecutingContext.ModelState);
-            var objectResult = new ObjectResult(apiProblemDetails);
-
-            objectResult.StatusCode = actionExecutingContext.HttpContext.Response.StatusCode;
+            var objectResult = new ObjectResult(apiProblemDetails)
+            {
+                StatusCode = actionExecutingContext.HttpContext.Response.StatusCode
+            };
             actionExecutingContext.Result = objectResult;
         }
 
         private void ExceptionApiProblemDetails(ActionExecutingContext actionExecutingContext, ActionExecutedContext actionExecutedContext, Exception ex)
         {
             var apiProblemDetails = new ApiProblemDetails(actionExecutingContext.HttpContext, ex);
-            var objectResult = new ObjectResult(apiProblemDetails);
-
-            objectResult.StatusCode = actionExecutingContext.HttpContext.Response.StatusCode;
+            var objectResult = new ObjectResult(apiProblemDetails)
+            {
+                StatusCode = actionExecutingContext.HttpContext.Response.StatusCode
+            };
             actionExecutedContext.Result = objectResult;
             actionExecutedContext.ExceptionHandled = true;
         }
 
-        private void AfterResultApiProblemDetails(ActionExecutedContext actionExecutedContext)
+        private void AfterResultApiProblemDetails(ActionExecutedContext actionExecutedContext, string title = default(string))
         {
-            var apiProblemDetails = new ApiProblemDetails(actionExecutedContext.HttpContext);
-            var objectResult = new ObjectResult(apiProblemDetails);
-
-            objectResult.StatusCode = actionExecutedContext.HttpContext.Response.StatusCode;
+            var apiProblemDetails = new ApiProblemDetails(actionExecutedContext.HttpContext, title);
+            var objectResult = new ObjectResult(apiProblemDetails)
+            {
+                StatusCode = actionExecutedContext.HttpContext.Response.StatusCode
+            };
             actionExecutedContext.Result = objectResult;
         }
     }
