+++ OAuth/appsettings.json
@@ -1,6 +1,6 @@
 {
   "AppOptions": {
-    "Secret": "THIS IS USED TO SIGN AND VERIFY JWT TOKENS, REPLACE IT WITH YOUR OWN SECRET, IT CAN BE ANY STRING"
+    "Secret": "EAAAAIEf3Z58KjJlC5S7gSHf2jKcW/c0mL3MID7AEnALxau+9MIh4sCZtCyyHEQgM0Qay0Qts5hDOT0BF9UpwUjUi/I="
   },
   "DBApiOptions": {
     "Url": "http://localhost:44335/api/"
@@ -8,7 +8,7 @@
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "Microsoft": "Warning",
+      "Microsoft": "Error",
       "Microsoft.Hosting.Lifetime": "Information"
     }
   }
